Update firefox config

+++ save_config.sh
@@ -36,6 +36,6 @@ pip freeze --user > pip/requirements.txt
 for profile in $(find $HOME/.mozilla/firefox -maxdepth 1 -mindepth 1 -type d); do
     profile_name="$(basename $profile)"
     mkdir -p "firefox/$profile_name"
-    cp -r "$profile/chrome" "firefox/$profile_name/chrome"
-    cp "$profile/user.js" "firefox/$profile_name/user.js"
+    [ -d "$profile/chrome" ] && cp -r "$profile/chrome" "firefox/$profile_name/chrome"
+    [ -f "$profile/user.js" ] && cp "$profile/user.js" "firefox/$profile_name/user.js"
 done
